# backend/public/docs/openapi.yaml
openapi: 3.0.3
info:
  title: 'Crypto Tracker API Documentation'
  description: 'API documentation for the Crypto Tracker application. This API allows users to track cryptocurrency prices, follow/unfollow cryptocurrencies, and get market data.'
  version: 1.0.0
servers:
  -
    url: 'https://backend.ddev.site'
...
components:
  securitySchemes:
    default:
      type: http
      scheme: bearer
      description: 'You can retrieve your token by visiting your dashboard and clicking <b>Generate API token</b>.'
security:
  -
    default: []
paths:
...
  /api/email/verification-notification:
    post:
      summary: 'Resend the email verification notification.'
      operationId: resendTheEmailVerificationNotification
      description: ''
      parameters: []
      responses:
        200:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  message: 'Verification link resent to your email address.'
                properties:
                  message:
                    type: string
                    example: 'Verification link resent to your email address.'
        400:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  message: 'Email already verified.'
                  is_verified: true
                properties:
                  message:
                    type: string
                    example: 'Email already verified.'
                  is_verified:
                    type: boolean
                    example: true
        401:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  message: Unauthenticated.
                properties:
                  message:
                    type: string
                    example: Unauthenticated.
      tags:
        - Authentication
...